refactor: move theme switch to component
