--- src/components/HelloWorld.vue
@@ -1,7 +1,0 @@
-<template>
-  {{ helloWorld }}
-</template>
-
-<script setup lang="ts">
-const helloWorld = greetings()
-</script>
